+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 28.03.2023 [v2.25.0]
+- Fixed bug with missing spaces in notes.
 - Updated Czech translation [Jerry Verner].
 - Added display of days remaining until birth anniversaries, multiples of 10 and 25 years.
 - Added display of marker tooltips with place names and dates in maps.